Create README.md
The API handles functionality for inserting and retrieving features specified in the models. In this case "Books", it also allows for update and delete using the PUT and DELETE methods through the book/details view. A simple /home/ url has been added using Html and Css to display the main features. It will easily store the data in the forms and added into the sqlite database. The select dropdown menu will be updated, displaying the items for this specific model for immediate retrieval. If the item is selected it will be displayed in the site's table.